requirements.txt
slither-analyzer
networkx